ETL/requirement.txt
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
google-api-core==1.30.0
google-auth==1.31.0
google-cloud-bigquery==2.20.0
google-cloud-core==1.6.0
google-cloud-datastore==2.1.3
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==2.10
libcst==0.3.19
mypy-extensions==0.4.3
numpy==1.20.3
packaging==20.9
pandas==1.2.4
proto-plus==1.18.1
protobuf==3.17.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.5
